chore: refactor code structure
