+++ packages/reactive/src/index.spec.ts
@@ -1,97 +1,12 @@
-import { describe, it, expect, vitest } from "vitest";
-import { renderHook } from "@testing-library/react";
+import { describe, it, expect } from "vitest";
 
-import { create } from "./index.js";
-import { getSnapshot } from "./snapshot.js";
-import { useSnapshot } from "./use-snapshot.js";
+import { ref, ProxyMap, ProxySet, create } from "./index";
 
-const runMacroTask = (fn: Function) => setTimeout(fn, 0);
-
-describe("index", () => {
-  it("create, proxy, useSnapshot and subscribe should be defined", () => {
+describe("index export", () => {
+  it("[ref, ProxyMap, ProxySet, create] should be defined", () => {
+    expect(ref).toBeDefined();
+    expect(ProxyMap).toBeDefined();
+    expect(ProxySet).toBeDefined();
     expect(create).toBeDefined();
-  });
-
-  it("should return mutate, useSnapshot and subscribe by current state", () => {
-    const store = create({
-      name: "Pikachu",
-      address: {
-        city: "NanJing",
-      },
-    });
-
-    expect(store.mutate).toMatchObject({
-      name: "Pikachu",
-      address: {
-        city: "NanJing",
-      },
-    });
-
-    const { result } = renderHook(() => useSnapshot(store.mutate));
-    expect(result.current).toEqual(getSnapshot(store.mutate));
-
-    const callback = vitest.fn();
-    store.subscribe(callback);
-
-    expect(callback).toHaveBeenCalledTimes(0);
-
-    store.mutate.name = "Charmander";
-    runMacroTask(() => expect(callback).toHaveBeenCalledTimes(1));
-  });
-
-  it("should restore state when restore function was called", () => {
-    const store = create({
-      name: "Pikachu",
-      address: {
-        city: "NanJing",
-      },
-    });
-
-    store.mutate.name = "Charmander";
-    store.restore();
-
-    runMacroTask(() =>
-      expect(store.mutate).toMatchObject({
-        name: "Pikachu",
-        address: {
-          city: "NanJing",
-        },
-      })
-    );
-  });
-
-  it("should subscribe and unsubscribe", () => {
-    const store = create({
-      name: "Pikachu",
-      address: {
-        city: "NanJing",
-      },
-    });
-
-    const callback = vitest.fn();
-    const unsubscribe = store.subscribe(callback);
-
-    store.mutate.name = "Charmander";
-    runMacroTask(() => expect(callback).toHaveBeenCalledTimes(1));
-
-    unsubscribe();
-    store.mutate.name = "Pikachu";
-    runMacroTask(() => expect(callback).toHaveBeenCalledTimes(1));
-  });
-
-  it("should notify when delete a property", () => {
-    const store = create({
-      name: "Pikachu",
-      address: {
-        city: "NanJing",
-      },
-    });
-
-    const callback = vitest.fn();
-    store.subscribe(callback);
-
-    delete store.mutate.name;
-
-    runMacroTask(() => expect(callback).toHaveBeenCalledTimes(1));
   });
 });
