+++ packages/reactive/src/vanilla/produce.ts
@@ -1,12 +1,10 @@
 import { proxy } from "./proxy.js";
 import { getSnapshot } from "./snapshot.js";
 
-import type { DeepReadonly } from "./internal-utils.js";
-
 export function produce<T extends object>(obj: T, handler: (draft: T) => void) {
   let state = proxy(obj);
   handler(state);
-  const snapshot = getSnapshot(state) as DeepReadonly<T>;
+  const snapshot = getSnapshot(state) as T;
   state = null;
   return snapshot;
 }
