+++ packages/reactive/src/vanilla/proxy-map.ts
@@ -1,0 +1,7 @@
+// TODO: need to implement
+
+export class ProxyMap {
+  constructor() {
+    return {};
+  }
+}
