+++ packages/reactive/src/vanilla/proxy-set.ts
@@ -1,0 +1,7 @@
+// TODO: need to implement
+
+export class ProxySet {
+  constructor() {
+    return {};
+  }
+}
