+++ packages/reactive/src/vanilla/snapshot.ts
@@ -1,4 +1,4 @@
-import { SNAPSHOT } from "./internal-utils.js";
+import { SNAPSHOT } from "../utils/index.js";
 
 export function getSnapshot<T extends object>(proxyState: T): T {
   return proxyState[SNAPSHOT];
